Fix all compiler and linker warnings
- Wsign-compare: cast container.size() to int in loop/comparison
  conditions across cParticleDrawer, cGame_logic, cParticle,
  cDrawManager, cMap, cIniGameRessouces, common, ini

- Wmismatched-tags: change forward declaration of s_PreviewMap in
  cEditorState.h from 'class' to 'struct' to match its definition

- Wunused-private-field: remove unused m_winFlags/m_loseFlags from
  cGamePlaying (shadowed duplicate of cGameConditionChecker fields),
  initialDuration from cPlayerNotification (only referenced in
  commented-out code), and numberFile from cPack

- Wunused-lambda-capture: remove unused 'this' from mentat yes-button
  lambdas (they use global 'game', not 'this'), remove dead qMaps
  page-count calculation from cSetupSkirmishState (the lambda that
  captured it never used it)

- Linker warning: deduplicate SDL libraries in CMakeLists.txt via
  list(REMOVE_DUPLICATES) to avoid 'ignoring duplicate libraries'
  from ld on macOS

# Conflicts:
#	src/drawers/cParticleDrawer.cpp
#	src/player/cPlayerNotification.cpp
#	src/player/cPlayerNotification.h

+++ src/game/cGame_logic.cpp
@@ -197,7 +197,7 @@ void cGame::init()
         m_Units[i].init(i);
     }
 
-    for (int i = 0; i < m_particles.size(); i++) {
+    for (int i = 0; i < (int)m_particles.size(); i++) {
         m_particles[i].init();
     }
