+++ src/managers/cDrawManager.cpp
@@ -123,7 +123,7 @@ void cDrawManager::drawDebugInfoUsages() const
     }
 
     int bulletsUsed = 0;
-    for (int i = 0; i < game.g_Bullets.size(); i++) {
+    for (int i = 0; i < (int)game.g_Bullets.size(); i++) {
         cBullet &pBullet = game.g_Bullets[i];
         if (pBullet.bAlive) {
             bulletsUsed++;
@@ -131,7 +131,7 @@ void cDrawManager::drawDebugInfoUsages() const
     }
 
     int particlesUsed = 0;
-    for (int i = 0; i < game.m_particles.size(); i++) {
+    for (int i = 0; i < (int)game.m_particles.size(); i++) {
         cParticle &pParticle = game.m_particles[i];
         if (pParticle.isValid()) {
             particlesUsed++;
